// frontendDio/front-prediction/src/app/interfaces/house.ts
export interface House {
	id : number | undefined;
	price: number;
	bedrooms: number;
	bathrooms: number;
	sqftLiving: number;
	sqftLot: number;
	floors: number;
	view: number;
	condition: number;
	grade: number;
	yearBuilt: number;
	yearRenovated: number;
	sqftAbove: number;
	sqftBasement: number;
}